+++ src/face_detection.ts
@@ -10,7 +10,7 @@ export class FaceDetection {
   private apiKey: string;
   private baseUrl: string;
 
-  constructor(apiKey: any, baseUrl: string = "https://sandbox.api.sap.com") {
+  constructor(apiKey: any, baseUrl: string = "https://sandbox.api.sap.com/ml/api/v2alpha1/image/face-detection") {
     assert(apiKey, "apiKey is required");
     this.apiKey = apiKey;
     this.baseUrl = baseUrl;
@@ -33,7 +33,7 @@ export class FaceDetection {
           Accept: "application/json",
         };
 
-        const url = this.baseUrl + "/ml/facedetection/face-detection";
+        const url = this.baseUrl;
 
         request.post({ url, formData, headers }, (err, response, body) => {
           if (err) {