+++ src/human_detection.ts
@@ -10,7 +10,7 @@ export class HumanDetection {
   private apiKey: string;
   private baseUrl: string;
 
-  constructor(apiKey: any, baseUrl: string = "https://sandbox.api.sap.com") {
+  constructor(apiKey: any, baseUrl: string = "https://sandbox.api.sap.com/ml/api/v2alpha1/image/human-detection/") {
     assert(apiKey, "apiKey is required");
     this.apiKey = apiKey;
     this.baseUrl = baseUrl;
@@ -33,7 +33,7 @@ export class HumanDetection {
           Accept: "application/json",
         };
 
-        const url = this.baseUrl + "/ml/humandetection/human-detection/";
+        const url = this.baseUrl;
 
         request.post({ url, formData, headers }, (err, response, body) => {
           if (err) {
@@ -62,7 +62,7 @@ export class HumanDetection {
           APIKey: this.apiKey,
         };
 
-        const url = this.baseUrl + "/ml/humandetection/human-detection/format:image";
+        const url = this.baseUrl + "format:image";
 
         const imageData: any = [];